seems that having padfstring receiving unsigned int fixes gfortran x64 problem
git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@9027 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ fwrap.cpp
@@ -25,7 +25,7 @@ f2cstring(char* fstring, size_t len)
 }
 
 void
-padfstring(char *dest, const char *src, size_t len)
+padfstring(char *dest, const char *src, unsigned int len)
 {
     size_t sofar;
